--- compressor/tests/test_compress_command.py
@@ -1,27 +1,0 @@
-from django.core.management import call_command
-from django.test import TestCase
-from compressor.cache import get_offline_manifest
-from .test_offline import OfflineTestCaseMixin
-
-
-class TestCompressCommand(OfflineTestCaseMixin, TestCase):
-    templates_dir = "test_compress_command"
-
-    def test_multiple_engines(self):
-        call_command('compress', force=True, engines=["django"])
-        manifest_django = get_offline_manifest()
-
-        call_command('compress', force=True, engines=["jinja2"])
-        manifest_jinja2 = get_offline_manifest()
-
-        call_command('compress', force=True, engines=["django", "jinja2"])
-        manifest_both = get_offline_manifest()
-
-        manifest_both_expected = {}
-        manifest_both_expected.update(manifest_django)
-        manifest_both_expected.update(manifest_jinja2)
-
-        self.assertNotEqual(manifest_django, manifest_jinja2)
-        self.assertTrue(manifest_django)
-        self.assertTrue(manifest_jinja2)
-        self.assertEqual(manifest_both, manifest_both_expected)
