+++ compressor/tests/test_offline.py
@@ -10,6 +10,7 @@ from importlib import import_module
 from mock import patch
 from unittest import SkipTest, skipIf
 
+from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.template import Template, Context
 from django.test import TestCase
@@ -655,3 +656,26 @@ class OfflineCompressExtendsRecursionTestCase(OfflineTestCaseMixin, TestCase):
         count, result = CompressCommand().compress(
             log=self.log, verbosity=self.verbosity, engine=engine)
         self.assertEqual(count, 1)
+
+
+class TestCompressCommand(OfflineTestCaseMixin, TestCase):
+    templates_dir = "test_compress_command"
+
+    def test_multiple_engines(self):
+        call_command('compress', force=True, engines=["django"])
+        manifest_django = get_offline_manifest()
+
+        call_command('compress', force=True, engines=["jinja2"])
+        manifest_jinja2 = get_offline_manifest()
+
+        call_command('compress', force=True, engines=["django", "jinja2"])
+        manifest_both = get_offline_manifest()
+
+        manifest_both_expected = {}
+        manifest_both_expected.update(manifest_django)
+        manifest_both_expected.update(manifest_jinja2)
+
+        self.assertNotEqual(manifest_django, manifest_jinja2)
+        self.assertTrue(manifest_django)
+        self.assertTrue(manifest_jinja2)
+        self.assertEqual(manifest_both, manifest_both_expected)
